// packages/sdk-router/src/typechain/FastBridgeV2.ts
export declare namespace IFastBridge {
  export type BridgeParamsStruct = {
    dstChainId: BigNumberish
    sender: string
    to: string
    originToken: string
    destToken: string
    originAmount: BigNumberish
    destAmount: BigNumberish
    sendChainGas: boolean
    deadline: BigNumberish
  }

  export type BridgeParamsStructOutput = [
    number,
    string,
    string,
    string,
    string,
    BigNumber,
    BigNumber,
    boolean,
    BigNumber
  ] & {
    dstChainId: number
    sender: string
    to: string
    originToken: string
    destToken: string
    originAmount: BigNumber
    destAmount: BigNumber
    sendChainGas: boolean
    deadline: BigNumber
  }

  export type BridgeTransactionStruct = {
    originChainId: BigNumberish
    destChainId: BigNumberish
    originSender: string
    destRecipient: string
    originToken: string
    destToken: string
    originAmount: BigNumberish
    destAmount: BigNumberish
    originFeeAmount: BigNumberish
    sendChainGas: boolean
    deadline: BigNumberish
    nonce: BigNumberish
  }

  export type BridgeTransactionStructOutput = [
    number,
    number,
    string,
    string,
    string,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    boolean,
    BigNumber,
    BigNumber
  ] & {
    originChainId: number
    destChainId: number
    originSender: string
    destRecipient: string
    originToken: string
    destToken: string
    originAmount: BigNumber
    destAmount: BigNumber
    originFeeAmount: BigNumber
    sendChainGas: boolean
    deadline: BigNumber
    nonce: BigNumber
  }
}